requirements.txt
flask>=2.3.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
openpyxl>=3.1.2
scikit-image>=0.21.0
scipy>=1.11.0
tensorflow>=2.13.0
pdf2image>=1.16.0
pypdf>=3.0.0
